+++ test/mocks/xml/index.ts
@@ -1,0 +1,4 @@
+export * from './context.xml'
+export * from './event.xml'
+export * from './mission.xml'
+export * from './resource.xml'